refactor(console): merge auth and shell store #2
