+++ apps/console/app/components/ui/ModelSelector.tsx
@@ -3,7 +3,7 @@ import { useSnapshot } from "valtio";
 import { Badge } from "@hebo/shared-ui/components/Badge";
 import { Select } from "@hebo/shared-ui/components/Select";
 
-import { shellStore } from "~console/state/auth";
+import { shellStore } from "~console/lib/shell";
 
 import type { ComponentProps } from "react";
