+++ apps/console/app/lib/auth/dummy-auth.ts
@@ -1,7 +1,7 @@
 import { Collection } from "@msw/data";
 import { z } from "zod";
 
-import { shellStore } from "~console/state/auth";
+import { shellStore } from "~console/lib/shell";
 
 import { DEFAULT_EXPIRATION_MS, type AuthService } from "./types";
