+++ apps/console/app/lib/shell.ts
